2Plane x 16KB..
G.C 동작 및 Host request 처리관련 path 수정함.
추후 DMA 시간관련 확인해야함.

+++ ftl/algo/abm.h
@@ -47,7 +47,8 @@ enum BDBM_ABM_SUBPAGE_STATUS {
 	BDBM_ABM_SUBPAGE_INVALID,
 };
 
-typedef uint8_t babm_abm_subpage_t; /* BDBM_ABM_PAGE_STATUS */
+//typedef uint8_t babm_abm_subpage_t; /* BDBM_ABM_PAGE_STATUS */
+typedef uint64_t babm_abm_subpage_t; /* BDBM_ABM_PAGE_STATUS */
 
 enum BDBM_ABM_BLK_STATUS {
 	BDBM_ABM_BLK_FREE = 0,
